% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeneCellCorr.R
\name{calculateFeatureCorrs}
\alias{calculateFeatureCorrs}
\title{genewise correlation}
\usage{
calculateFeatureCorrs(sim_list, ncore = 8)
}
\arguments{
\item{sim_list}{a list containing the real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
A data frame with genewise correlations for real and simulated data
}
\description{
genewise correlation
}
