% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanVarLibrary.R
\name{calculateMeanVarLibrary}
\alias{calculateMeanVarLibrary}
\title{Calculate mean variance and library size estimates}
\usage{
calculateMeanVarLibrary(sim_list, ncore = 8)
}
\arguments{
\item{sim_list}{A list containing real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
a list containing the mean , variance , library size estimates of the real and simulated data for downstream evaluation
}
\description{
Calculate mean variance and library size estimates
}
