% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeneCellCorr.R
\name{calculateSampleCorrs}
\alias{calculateSampleCorrs}
\title{cellwise correlation}
\usage{
calculateSampleCorrs(sim_list, maxNForCorr = 1000, ncore = 8)
}
\arguments{
\item{sim_list}{a list containing the real and simulated data}

\item{maxNForCorr}{number of cells selected to compute the cellwise correlation}

\item{ncore}{number of cores for parallel computing}
}
\value{
A data frame with cellwise correlations for real and simulated data
}
\description{
cellwise correlation
}
