% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsim_eval_KS.R
\name{countsim_eval}
\alias{countsim_eval}
\title{Calculate KDE test statistic on each parameter}
\usage{
countsim_eval(sim_list, ncore = 8)
}
\arguments{
\item{sim_list}{A list containing real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
the KDE test statistic across 13 parameters, and the raw values that are used to calculate the KDE test statistics
}
\description{
Calculate KDE test statistic on each parameter
}
