% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_biosignal_plot}
\alias{draw_biosignal_plot}
\title{Plot biological plot}
\usage{
draw_biosignal_plot(result)
}
\arguments{
\item{sim_list}{A list containing real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
the KDE test statistic across 13 parameters, and the raw values that are used to calculate the KDE test statistics
}
\description{
Plot biological plot
}
