% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{eval_parameter}
\alias{eval_parameter}
\title{run evaluation on parameter estimation}
\usage{
eval_parameter(real, sim, type = "count", method = "samplemethod")
}
\arguments{
\item{real}{real data}

\item{sim}{simulated data}

\item{type}{whether the simulated data is normalised or unnormalised}

\item{method}{name of the simulation method}
}
\value{
distribution of parameters that are used to construct KDE test, KDE test statistics of each cell type, combined KDE test statistics
}
\description{
run evaluation on parameter estimation
}
