% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_signal_prop.R
\name{eval_signal}
\alias{eval_signal}
\title{evaluate the similarities between the two set of DE genes}
\usage{
eval_signal(real, sim)
}
\arguments{
\item{real}{real data}

\item{simulated}{simulated data}
}
\value{
confusion matrix, logFC of the set of genes from real and simiulated data
}
\description{
evaluate the similarities between the two set of DE genes
}
