% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DE_prop.R
\name{evaluate_DE}
\alias{evaluate_DE}
\title{Evaluate DE}
\usage{
evaluate_DE(real, simulated)
}
\arguments{
\item{real}{A list containing real and simulated data}

\item{simulated}{number of cores for parallel computing}
}
\value{
DE
}
\description{
Evaluate DE
}
