% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DE_prop.R
\name{generate_DE}
\alias{generate_DE}
\title{Generate DE}
\usage{
generate_DE(
  exprsMat,
  trainClass,
  feature = c("DE", "DV", "DD", "DP", "BD"),
  pSig = 0.1
)
}
\arguments{
\item{sim_list}{A list containing real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
the KDE test statistic across 13 parameters, and the raw values that are used to calculate the KDE test statistics
}
\description{
Generate DE
}
