% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_test.R
\name{KDE_test}
\alias{KDE_test}
\title{Perform KDE test}
\usage{
KDE_test(df, column, subsampleSize)
}
\arguments{
\item{df}{dataframe containing the raw distribution for comparison}

\item{column}{the column containing the parameter of interest}

\item{subsampleSize}{maximum number of data points for comparison}
}
\value{
KDE test statistic
}
\description{
Perform KDE test
}
