% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{prepare_eval}
\alias{prepare_eval}
\title{format the data for evaluation}
\usage{
prepare_eval(sim_list, thissim, name, celltype, counttype)
}
\arguments{
\item{sim_list}{a list for storing the real and simulated data}

\item{thissim}{the data to be put into the sim_list}

\item{name}{name of the data (eg, "real", "simulated")}

\item{celltype}{cell type that the data contains}

\item{counttype}{whether the data is unnormalised or normalised}
}
\value{
a sim_list filled with the formatted data
}
\description{
format the data for evaluation
}
