% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DE_prop.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(real, simulated)
}
\arguments{
\item{sim_list}{A list containing real and simulated data}

\item{ncore}{number of cores for parallel computing}
}
\value{
the KDE test statistic across 13 parameters, and the raw values that are used to calculate the KDE test statistics
}
\description{
Process data
}
