% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{summarise_score}
\alias{summarise_score}
\title{summarise the scores across multiple cell type according to the proportion of each cell type}
\usage{
summarise_score(intermediate)
}
\arguments{
\item{intermediate}{the scores from multiple cell type}
}
\value{
a summarised score
}
\description{
summarise the scores across multiple cell type according to the proportion of each cell type
}
